refactor: consolidate conversion utility into single standalone script
- Remove conversion module from openhands.sdk.context.skills.conversion
- Remove tests for conversion module
- Make scripts/convert_legacy_skills.py a standalone script with all logic included
- Script no longer depends on SDK module, only requires python-frontmatter
