--- openhands-sdk/openhands/sdk/context/skills/conversion/__init__.py
@@ -1,18 +1,0 @@
-"""Conversion utilities for legacy OpenHands skills to AgentSkills format."""
-
-from openhands.sdk.context.skills.conversion.converter import (
-    convert_legacy_skill,
-    convert_skills_directory,
-    generate_description,
-    normalize_skill_name,
-    validate_skill_name,
-)
-
-
-__all__ = [
-    "convert_legacy_skill",
-    "convert_skills_directory",
-    "generate_description",
-    "normalize_skill_name",
-    "validate_skill_name",
-]
